Hide UnpaddedTitle by making it private
